fix intermittent parse error with 5.19 perl
The test occasionally fail on their own due to this.  I'm not sure if
it's limited to 5.19 or if it can happen in older perls, but I found it
while looking for hash-related perl5.17+ bugs.

Symptom:  in Net::ISC::DHCPd if you run the leases test you'll get a
validation error about every 3rd run.  This is caused by the binding
state being 'free; set leasetime = blah'

In the POE::Filter tests you'll occasionally see it not find a mac
address when parsing the data.  This may vary depending on hardware.
The fix is to make the regex non-greedy.  In two cases where we know
that no spaces are allowed in the results I've changed the regex to \S+.