Changed the parser to line-by-line
This allows us to skip unknown values as long as they aren't inside a
lease statement.

I also changed the parser around for speed.  It tries to only do a regex
compare if an input line is possible. (should've done this in another
commit but I was having fun and forgot to commit stuff)

fixes #2